Revert "Merge branch 'develop' of https://github.com/733737/OpenRCT2 into develop"
This reverts commit 3d840bd9cfc449fc9a12d65b01f392fb6311aee2, reversing
changes made to 185a0e548359af09b6ab7592dabee6ef094d43ea.

+++ src/openrct2/world/Park.cpp
@@ -379,7 +379,7 @@ int32_t Park::CalculateParkRating() const
     // Guests
     {
         // -150 to +3 based on a range of guests from 0 to 2000
-        result -= 150 - (std::min<int32_t>(2000, gNumGuestsInPark) / 13);
+        result -= 150 - (std::min<int16_t>(2000, gNumGuestsInPark) / 13);
 
         // Find the number of happy peeps and the number of peeps who can't find the park exit
         uint32_t happyGuestCount = 0;
@@ -459,8 +459,8 @@ int32_t Park::CalculateParkRating() const
             result += 100 - averageExcitement - averageIntensity;
         }
 
-        totalRideExcitement = std::min<int32_t>(1000, totalRideExcitement);
-        totalRideIntensity = std::min<int32_t>(1000, totalRideIntensity);
+        totalRideExcitement = std::min<int16_t>(1000, totalRideExcitement);
+        totalRideIntensity = std::min<int16_t>(1000, totalRideIntensity);
         result -= 200 - ((totalRideExcitement + totalRideIntensity) / 10);
     }
